Fix small issues with tooltips
If a button was pressed that deleted tooltips, some of the buttons that didn't delete them didn't have functions to re-enable tooltips. Instead of going through every single button to add the add and remove tooltip functions I changed it around to something simpler that works more effectively in the long term.